--- src/main/java/com/codeup/capstonestarter/data/Type.java
@@ -1,44 +1,0 @@
-package com.codeup.capstonestarter.data;
-
-import javax.persistence.*;
-import java.util.Collection;
-
-@Entity
-@Table (name = "types")
-public class Type {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String type;
-
-    public Type() {
-    }
-    @OneToMany(mappedBy = "type")
-    private Collection<Event> events;
-
-    public Type(Long id, String type) {
-        this.id = id;
-        this.type = type;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-
-}
