+++ src/main/java/com/codeup/capstonestarter/data/event/EventRepository.java
@@ -1,4 +1,4 @@
-package com.codeup.capstonestarter.data;
+package com.codeup.capstonestarter.data.event;
 
 import org.springframework.data.jpa.repository.JpaRepository;
